# Conflicts:
#	platform/core-ui/src/util/IconUtil.kt

+++ platform/platform-impl/src/com/intellij/openapi/vfs/newvfs/persistent/PersistentFSRecordAccessor.java
@@ -1,8 +1,9 @@
-// Copyright 2000-2022 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
+// Copyright 2000-2023 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
 package com.intellij.openapi.vfs.newvfs.persistent;
 
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.diagnostic.Logger;
+import com.intellij.openapi.vfs.newvfs.persistent.FSRecordsImpl.FileIdIndexedStorage;
 import com.intellij.util.BitUtil;
 import it.unimi.dsi.fastutil.ints.IntArrayList;
 import it.unimi.dsi.fastutil.ints.IntList;
@@ -56,15 +57,39 @@ public final class PersistentFSRecordAccessor {
     connection.markDirty();
   }
 
-  public int createRecord() throws IOException {
+  public int createRecord(Iterable<FileIdIndexedStorage> fileIdIndexedStorages) throws IOException {
     connection.markDirty();
+
+    if (!FSRecordsImpl.REUSE_DELETED_FILE_IDS) {
+      return connection.getRecords().allocateRecord();
+    }
 
     final int reusedRecordId = connection.reserveFreeRecord();
     if (reusedRecordId < 0) {
       return connection.getRecords().allocateRecord();
     }
     else {//there is a record for re-use, but let's clean it up first:
+      //TODO RC: Actually, it could significantly slow down new record allocation, so it is better to clear all
+      //         that async, during free records collection.
+      //         We can do that for attributes & content, but we can't do it for fileIdIndexedStorages, since we
+      //         don't know the list of FileIdIndexedStorage storages early on -- fileIdIndexedStorages are collected
+      //         incrementally, as they are registered.
+      //         Actually, by the same reason even current approach is not bulletproof: since .createRecord() could
+      //         be called _before_ all FileIdIndexedStorage are registered, hence reusedRecordId in not-yet-registered
+      //         FileIdIndexedStorage won't be cleaned.
+      //         Alternative solutions could be:
+      //         a) Maintain (i.e. persist) list of 'associated storages' inside VFS, and read all them on startup.
+      //            Need additional fields in storage header (bytesPerRow) to read the storage content.
+      //         b) Clean data associated with removed fileIds on shutdown (when all storages are already registered),
+      //            not on startup. Delays shutdown, and increase chance of VFS corruption if app forcibly terminated
+      //            due to long shutdown.
+      //         c) Don't reuse fileId at all -- just keep allocating new fileIds, and re-build VFS as int32 exhausted
+      //            Actually, quite interesting approach, now under REUSE_DELETED_FILE_IDS feature-flag, but needs careful examination
+      //            (for the next release?)
       deleteContentAndAttributes(reusedRecordId);
+      for (FileIdIndexedStorage storage : fileIdIndexedStorages) {
+        storage.clear(reusedRecordId);
+      }
       connection.getRecords().cleanRecord(reusedRecordId);
       return reusedRecordId;
     }
